Remove the process for customBackend.json & grant testing gallery permission
Automation will be executed on staging backend thanks to the
installation of TestingGallery.

Thus no need to create and push customBackend.json to device when
TestingGallery installed.

BUT, if you want to change the backend by using customBackend.json, you
have to uninstall testing gallery otherwise it will not work.